requirements.txt
# Python 依赖

# 核心深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.6.0
PyWavelets>=1.1.0

# 数值计算
numpy>=1.19.0
scikit-image>=0.18.0

# 配置管理
PyYAML>=5.4.0

# 日志与可视化
tensorboard>=2.4.0
tensorboardX>=2.5
matplotlib>=3.3.0
tqdm>=4.60.0

# 工具
einops>=0.4.0
